Update output dtype and add requirements file
Changed the default `output_dtype` in `local_histogram_match` function from 'float32' to 'float16' for memory efficiency. Added a new `requirements.txt` file listing necessary dependencies.

+++ local_histogram_match.py
@@ -564,7 +564,7 @@ def local_histogram_match(
         calculation_dtype_precision = 'float32',
         floor_value: Optional[float] = None,
         gamma_bounds: Optional[Tuple[float, float]] = None,
-        output_dtype = 'float32',
+        output_dtype = 'float16',
         projection: str = "EPSG:4326",
         debug_mode: bool = False,
         ):
